将id改成int64

+++ common/act/execution.go
@@ -15,7 +15,7 @@ import (
 type Execution struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 	// 流程实例id
 	ProcInstID int64 `json:"proc_inst_id,omitempty"`
 	// 流程定义id
@@ -63,7 +63,7 @@ func (e *Execution) assignValues(columns []string, values []interface{}) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			e.ID = int(value.Int64)
+			e.ID = value.Int64
 		case execution.FieldProcInstID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field proc_inst_id", values[i])
